fix(monitoring): M1-M3+L1-L2 修复 + review fixes H1/M1-M4/L1-L11
监控模块全栈修复,涵盖 monitoring API 测试报告 docs/MONITORING_API_TEST_REPORT_2026-06-08.md
中识别的 5 个问题,以及两轮 code review 的全部 follow-up。

**核心修复**:
- M1: management 监控 API 路径去 /admin 前缀 (前端 6 处)
- M2: MyBatis SqlTimingInterceptor 接入 MetricsCollector,queryCount 从 0 → 真实计数
- M3: 引入 spring-boot-starter-actuator + micrometer-prometheus,4 个 actuator 端点
- L1: Caffeine L1 + Redis L2 组合缓存,5 个监控方法 @Cacheable
- L2: /actuator/prometheus OpenMetrics 导出

**Review Round 1 修复**:
- H1: TS interface 6 个对齐后端 DTO,MonitoringView 5 处字段同步
- M1: prometheus 部署 doc (nginx IP allowlist)
- M2: SqlTimingInterceptor 补 queryCursor @Signature
- M3: DatabaseStatsVO.slowQueries int→long
- M4: plugin test 真正驱动 wrapped Executor
- L1-L9: 7 个小清理 (javadoc, @Deprecated since, 删冗余等)

**Review Round 2 修复**:
- L10: getAllStats 用 safe<T> 替代 as T 断言,Promise.all + try/catch
- L11: i18n 翻译键 javaVersion/heapMax/nonHeapUsed (en + zh-CN)

**新增**:
- common/metrics/MetricsCollector.java + SqlTimingInterceptor.java
- 测试 5+5 = 10 个新用例,30/30 tests green
- 端到端验证: 6 monitoring + 4 actuator 端点全 200

Refs: .claude/PRPs/reports/cosmic-chasing-eclipse-report.md
      .claude/reviews/fix-monitoring-issues-local-review.md
      .claude/reviews/fix-monitoring-issues-local-review-v2.md

+++ backend-spring/src/main/java/com/ulticode/modules/monitoring/dto/DatabaseStatsVO.java
@@ -29,9 +29,11 @@ public class DatabaseStatsVO {
     private Long queryCount;
 
     /**
-     * Number of slow queries detected.
+     * Number of slow queries detected since process start.
+     * Stored as {@code long} to avoid overflow on long-running processes
+     * (a busy OJ can easily produce > 2 billion slow queries in a year).
      */
-    private Integer slowQueries;
+    private Long slowQueries;
 
     /**
      * Database connection status.
